Reformatting GIFTs app
The GIFTs service now has the same formatting as other Production
services, with a navbar along the top, and defaulting to the
submission page. The jobs table needs much less complexity than
other services, so it has a somewhat stripped-down feel.

Also changed the colour scheme, green was (appropriately) bagged
by the Plant services, but nothing is using orange...